KA-160 Extract initial state to separate file
